Move the four DYNO screens out of EcuLab.jsx
Markup moved verbatim (inline styles intact) — CSS module conversion is a
separate commit so a failure has one candidate cause, not two.

DYNO's gating is irregular on purpose, unlike TUNE's four uniform
`dynoView === x` checks: while a pull is running, the switcher is hidden and
CURVES is the only view that can show, regardless of which section the URL
has selected. All five original conditions (the switcher's `!running`, CURVES'
`running || dynoView === 'result'`, and the other three's
`!running && dynoView === '<id>'`, SCORE also gated on `scores`) are preserved
exactly in the shell; only the per-section markup moved into screens/dyno/.

`buildHistogram`/`applyHistogram` moved into DataScreen.jsx (their only
caller); `dynoChartMaxRpm` moved into ResultScreen.jsx, computed off the
`engineDerived` prop, since neither has a second reader. `chartData`,
`engineDerived` and `scores` stay shell-owned props (each feeds a second
consumer — EcuScreen's fuel-trim chart, the tach/header, and DASH's
StatsScreen respectively); `bestScore` and everything else DYNO's screens
need is read straight off the store.

tests/theme.test.js's pull-log tone-classification scan re-points at
LogScreen.jsx, where that logic now lives — re-pointed, not relaxed, same as
button-call-sites.test.jsx's glob after each earlier tab's split.

tests/ui/dyno-screens.test.jsx adds the four screens mounted standalone (the
same property build/tune/dash-screens.test.jsx already pin for their tabs)
plus the regression this task exists to catch: selecting DATALOG and then
starting a second pull must fall back to the live curves, not go blank —
ROUTES.dyno[0] being 'result' means a naively normalised
`dynoView === 'result'` gate would pass every other existing test.
